==> _data_management/2b_extract_CMIP_historical.ipynb <==
# Google Cloud CMIP6 Public Data: Basic Python Example

https://gallery.pangeo.io/repos/pangeo-gallery/cmip6/basic_search_and_load.html

This notebooks shows how to query the catalog and load the data using python

The columns of the dataframe correspond to the CMI6 controlled vocabulary. A beginners' guide to these terms is available in [this document](https://docs.google.com/document/d/1yUx6jr9EdedCOLd--CPdTfGDwEwzPpCF6p1jRmqx-0Q). 

from matplotlib import pyplot as plt
import numpy as np
import pandas as pd
import xarray as xr
import zarr
import fsspec
import geopandas as gpd
import os

%matplotlib inline
%config InlineBackend.figure_format = 'retina' 
plt.rcParams['figure.figsize'] = 12, 6

data_dir = '/Users/maoyabassiouni/Documents/DATA/Projects/OLNI'

def get_df_CMIP(df_catalogue, query_string, print_q=None):
    df_q = df_catalogue.query(query_string)
    print(query_string, len(df_q.index))
    print('')
    if print_q:
        print(df_q[['source_id', 'experiment_id', 'table_id', 'variable_id', 'version', 'zstore']])
        print(list(set(df_q['source_id'].values)))
    
    # get the path to a specific zarr store (latest version if more than one)
    zstore = df_q.zstore.values[-1]
    print(zstore)
    print('')
    # create a mutable-mapping-style interface to the store
    mapper = fsspec.get_mapper(zstore)

    # open it using xarray and zarr
    ds = xr.open_zarr(mapper, consolidated=True)
    ds
    return ds

def make_loc_df(dsi, plot_locs, variable_name):
    [variable_id, experiment_id, source_id] = variable_name.split('_')
    data = np.array([])
    time_f = dsi.time.values#.astype("float64")
    lat_i, lon_i, sp_i = plot_locs[0]
    vals = dsi[variable_id].sel(lat=lat_i, lon =lon_i, method='nearest').values.astype("float64")
    ll = len(vals)
    print(lat_i, lon_i, sp_i, ll)
    data = np.vstack([np.ones(ll) * lat_i, 
               np.ones(ll) * lon_i, 
               np.ones(ll) * sp_i, 
               vals]).T 
    time_ff = time_f 
    for lat_i, lon_i, sp_i in plot_locs[1:]:
        vals = dsi[variable_id].sel(lat=lat_i, lon =lon_i, method='nearest').values.astype("float64")
        ll = len(vals)
        #print(lat_i, lon_i, sp_i, ll)
        data_i = np.vstack([np.ones(ll) * lat_i, 
                   np.ones(ll) * lon_i, 
                   np.ones(ll) * sp_i, 
                   vals]).T 
        data = np.concatenate([data, data_i])
        time_ff = np.concatenate([time_ff, time_f])
    lat, lon, sp, values = zip(*data)
    df_v = pd.DataFrame({'Lat': lat, 'Lon': lon, 'time': time_ff, 'sp': sp, variable_name: values})

    return df_v

def make_loc_df_X(dsi, plot_locs, variable_name):
    [variable_id, experiment_id, source_id] = variable_name.split('_')
    data = np.array([])
    time = dsi.time.values
    
    lat_i, lon_i, sp_i = plot_locs[0]
    vals = dsi[variable_id].sel(lat=lat_i, lon =lon_i, method='nearest').values
    ll = len(vals)
    #print(lat_i, lon_i, sp_i, ll)
    data = np.vstack([np.ones(ll) * lat_i, 
               np.ones(ll) * lon_i, 
               np.ones(ll) * sp_i, 
               time, vals]).T 
        
    for lat_i, lon_i, sp_i in plot_locs[1:]:
        vals = dsi[variable_id].sel(lat=lat_i, lon =lon_i, method='nearest').values
        ll = len(vals)
        #print(lat_i, lon_i, sp_i, ll)
        data_i = np.vstack([np.ones(ll) * lat_i, 
                   np.ones(ll) * lon_i, 
                   np.ones(ll) * sp_i, 
                   time, vals]).T 
        data = np.concatenate([data, data_i])
    lat, lon, sp, time, values = zip(*data)
    
    df_v = pd.DataFrame({'Lat': lat, 'Lon': lon, 'datetime': time, 'sp': sp, variable_name: values})
    return df_v

df_catalogue = pd.read_csv('https://storage.googleapis.com/cmip6/cmip6-zarr-consolidated-stores.csv')
df_catalogue.head()

scenarios = [['CMIP', 'historical'], ['ScenarioMIP', 'ssp245'],['ScenarioMIP', 'ssp585']]

variables_A = [['tas', 'Amon'], # Near-Surface Air Temperature [K]
             ['hurs', 'Amon'], #Near-Surface Relative Humidity [%]
             ['pr', 'Amon'], #Precipitation [kg m-2 s-1]
             ['rsds', 'Amon'], #Surface Downwelling Shortwave Radiation [W m-2]
             ]

variables_B = [['evspsblpot', 'Emon'], # Potential Evapotranspiration [kg m-2 s-1]
             ['co2mass','Amon'], #Total Atmospheric Mass of CO2 [kg]
             ['co2s', 'Emon'],  # Atmosphere CO2 [1e-06]
             ]

## finding models in catalogue with required variables

print('variables', variables_A)
print('')

models_set_all = []
for variable_id, table_id in variables_A:
    print(variable_id, table_id)
    models_v = []
    for activity_id, experiment_id in scenarios:
        query_string = "activity_id== '%s' & table_id == '%s' & variable_id == '%s' \
                        & experiment_id == '%s'"  \
                        % (activity_id, table_id, variable_id, experiment_id)

        df_q = df_catalogue.query(query_string)
        models = list(set(df_q['source_id'].values))
        models_v.append(models)
        print('\t', experiment_id, len(models))
        #print(models[:6])
    models_set = []
    for m in models_v[0]:
 
        mi = [1 for i in models_v[1:] if m in i]
        if len(mi)==(len(scenarios)-1):
            models_set.append([m, len(mi)+1])
    models_set_all.append(models_set)
    print('n set models', len(models_set))
    print('')
    
models_select = []
for m, nm in models_set_all[0]:
    mi = [1 for i in models_set_all[1:] if m in list(zip(*i))[0]]
    if len(mi)==(len(variables_A)-1):
        models_select.append([m, len(mi)+1])
print('Selected models', len(models_select))
print(list(zip(*models_select))[0])


models_set_all = []
for variable_id, table_id in variables_B:
    print(variable_id, table_id)
    models_v = []
    for activity_id, experiment_id in scenarios:
        query_string = "activity_id== '%s' & table_id == '%s' & variable_id == '%s' \
                        & experiment_id == '%s'"  \
                        % (activity_id, table_id, variable_id, experiment_id)

        df_q = df_catalogue.query(query_string)
        models = list(set(df_q['source_id'].values))
        models_v.append(models)
        print('\t', experiment_id, len(models))
        #print(models[:6])
    models_set = []
    for m in models_v[0]:
 
        mi = [1 for i in models_v[1:] if m in i]
        if len(mi)==(len(scenarios)-1):
            models_set.append([m, len(mi)+1])
    models_set_all.append(models_set)
    print('n set models', len(models_set))
    print(models_set)
    

## Extract loc data

df_ts_ = pd.read_csv(os.path.join(data_dir ,'model_inputs', 'df_model_all_annual.csv'))
df_ts_ = df_ts_[(df_ts_['yrange']>=10) & (df_ts_['count']>=6) 
                & (df_ts_['MODIS_Lai'].notna()) & (df_ts_['SATPSI'].notna())
                & (df_ts_['LNC'].notna())]
df_ts_ = df_ts_[(df_ts_['year']>=1995) & (df_ts_['year']<=2016)]

plot_locs = df_ts_.groupby(['Lat', 'Lon', 'sp']).mean().index
plot_locs

df_area = df_catalogue.query("variable_id == 'areacella' & source_id == 'GFDL-ESM4'")
ds_area = xr.open_zarr(fsspec.get_mapper(df_area.zstore.values[0]), consolidated=True)
plot_locs_cmip = []
for lat_i, lon_i, sp_i in plot_locs:
    ds_area_i = ds_area.sel(lat=lat_i, lon =lon_i, method='nearest')
    lon_c = ds_area_i['lon'].values.item()  
    lat_c = ds_area_i['lat'].values.item()  
    #lon_c = ds_area.lon.sel(lat=lat_i, lon =lon_i, method='nearest').values
    plot_locs_cmip.append('%5.4f_%5.4f_%d' %(lat_c, lon_c, sp_i))

print(len(plot_locs_cmip))
plot_locs_cmip_set = list(set(plot_locs_cmip))
plot_locs_cmip_set = [loc.split('_') for loc in plot_locs_cmip_set]
plot_locs_cmip_set = [[float(l) for l in loc] for loc in plot_locs_cmip_set]
print(len(plot_locs_cmip_set))

## Extract climate timeseries by model

models = [##'ACCESS-CM2',
          ##'CanESM5',
          ##'CESM2',
          #'CESM2-WACCM',
          #'CMCC-CM2-SR5',
          ##'CNRM-CM6-1',
          #'CNRM-CM6-1-HR',
          #'CNRM-ESM2-1',
          #'CanESM5-CanOE',
          #'EC-Earth3',
          #'EC-Earth3-Veg',
          ##'EC-Earth3-Veg-LR',
          #'FGOALS-f3-L',
          ##'FGOALS-g3',
          #'FIO-ESM-2-0',
          ##'GFDL-CM4',
          #'GFDL-ESM4',
          ##'GISS-E2-1-G',
          #'HadGEM3-GC31-LL',
          'IITM-ESM',
          #'INM-CM4-8',
          'INM-CM5-0',
          'IPSL-CM6A-LR',
          #'KACE-1-0-G',
          'KIOST-ESM',
          #'MIROC-ES2L',
          'MIROC6',
          #'MPI-ESM1-2-HR',
          'MPI-ESM1-2-LR',
          'MRI-ESM2-0',
          'NorESM2-MM',
          'UKESM1-0-LL']

models = ['ACCESS-CM2',
          'CanESM5',
          'CESM2',
          'CNRM-CM6-1',
          'EC-Earth3-Veg-LR',
          'FGOALS-g3',
          'GFDL-CM4',
          'GISS-E2-1-G',
          'IITM-ESM',
          'INM-CM5-0',
          'IPSL-CM6A-LR',
          'KIOST-ESM',
          'MIROC6',
          'MPI-ESM1-2-LR',
          'MRI-ESM2-0',
          'NorESM2-MM',
          'UKESM1-0-LL']

variables_A = variables_A[:3]
variables_A

#source_id = 'CESM2'
locations = plot_locs
#locations = plot_locs_cmip_set
#scenarios = [['CMIP', 'historical'], ['ScenarioMIP', 'ssp245'],['ScenarioMIP', 'ssp585']]
activity_id = 'CMIP'
experiment_id = 'historical'
#for activity_id, experiment_id in scenarios:
for source_id in models:
    for variable_id, table_id in variables_A:
        variable_name = '%s_%s_%s' % (variable_id, experiment_id, source_id)
        print(variable_name, '.....................................................')
        print('.............................................................................................')
        query_string = "activity_id== '%s' & table_id == '%s' & variable_id == '%s' \
                        & experiment_id == '%s' & source_id == '%s'" \
                        % (activity_id, table_id, variable_id, experiment_id, source_id)

        dsi = get_df_CMIP(df_catalogue, query_string)
        df_v = make_loc_df(dsi, locations, variable_name)
        df_v.to_csv(os.path.join(data_dir, 'grid_data_extract/CMIP', '%s.csv' % (variable_name))) 


## Extract CO2 timeseries

min_lat = np.min(list(zip(*plot_locs_cmip_set))[0])
max_lat = np.max(list(zip(*plot_locs_cmip_set))[0])
min_lon = np.min(list(zip(*plot_locs_cmip_set))[1])
max_lon = np.max(list(zip(*plot_locs_cmip_set))[1])

df_area = df_catalogue.query("variable_id == 'areacella' & source_id == 'GFDL-ESM4'")
ds_area = xr.open_zarr(fsspec.get_mapper(df_area.zstore.values[0]), consolidated=True)
ds_area = ds_area.sel(lat=slice(min_lat, max_lat), lon=slice(min_lon, max_lon))
total_area = ds_area.areacella.sum(dim=['lon', 'lat'])

source_id = 'GFDL-ESM4'
table_id = 'Emon'
variable_id = 'co2s'
for activity_id, experiment_id in scenarios: 
    variable_name = '%s_%s_%s' % (variable_id, experiment_id, source_id)
    print(variable_name)
    query_string = "activity_id== '%s' & table_id == '%s' & variable_id == '%s' \
                            & experiment_id == '%s' & source_id == '%s'" \
                            % (activity_id, table_id, variable_id, experiment_id, source_id)

    dsi = get_df_CMIP(df_catalogue, query_string)
    dsi = dsi.sel(lat=slice(min_lat, max_lat), lon=slice(min_lon, max_lon))
    co2_timeseries = (dsi[variable_id] * ds_area.areacella).sum(dim=['lon', 'lat']) / total_area
    
    co2_y_timeseries = co2_timeseries.resample(time='AS').mean('time')
    co2_df = pd.DataFrame({'datetime': co2_y_timeseries.time, 'CO2_%s' % experiment_id: co2_y_timeseries.values})
    co2_df.to_csv(os.path.join(data_dir, 'grid_data_extract/CMIP', '%s.csv' % (variable_name))) 

df_area = df_catalogue.query("variable_id == 'areacella' & source_id == 'CESM2'")
ds_area = xr.open_zarr(fsspec.get_mapper(df_area.zstore.values[0]), consolidated=True)

query_string = "activity_id=='ScenarioMIP' \
                & table_id == 'Amon' \
                & variable_id == 'tas' \
                & experiment_id == 'ssp245'\
                & institution_id == 'NCAR'"
variable_id = 'tas'
dsi = get_df_CMIP(df_catalogue, query_string, print_q=True)

dsi[variable_id].sel(time='2050-01').squeeze().plot()

total_area = ds_area.areacella.sum(dim=['lon', 'lat'])
total_area

ta_timeseries = (dsi[variable_id] * ds_area.areacella).sum(dim=['lon', 'lat']) / total_area
ta_timeseries

%time ta_timeseries.load()
ta_timeseries.plot(label='monthly')
ta_timeseries.rolling(time=12).mean().plot(label='12 month rolling mean')
plt.legend()
