==> _data_management/3b_combine_compile_annual_growseas_data_CMIP.ipynb <==
import os
import sys

import pandas as pd
import numpy as np
from datetime import datetime as dt
import calendar
import matplotlib.pyplot as plt

### To do
average also for annual with only months T>0C

%load_ext autoreload
%autoreload 2

import warnings
warnings.filterwarnings('ignore')

data_dir = '/Users/maoyabassiouni/Documents/DATA/Projects/OLNI'

# Combine annual files

models = ['ACCESS-CM2',
          'CanESM5',
          'CESM2',
          'CNRM-CM6-1',
          'EC-Earth3-Veg-LR',
          'FGOALS-g3',
          'GFDL-CM4',
          'GISS-E2-1-G',
          'IITM-ESM',
          'INM-CM5-0',
          'IPSL-CM6A-LR',
          'KIOST-ESM',
          'MIROC6',
          'MPI-ESM1-2-LR',
          'MRI-ESM2-0',
          'NorESM2-MM',
          'UKESM1-0-LL']

scenarios = [ 'historical', 'ssp245',  'ssp585']

variables_A = ['tas', 'hurs', 'pr', 'rsds']

source_id = models[0]
scenario = scenarios[0]
ff = os.path.join(data_dir, 'grid_data_extract', 'CMIP', '%s_%s_%s.csv' % (variables_A[0], scenario, source_id))
df = pd.read_csv(ff)
df['timestamp'] = [dt.strptime(dti, '%Y-%m-%d %H:%M:%S')
                                 for dti in df['datetime'].values]
df['month'] = [dti.astype('datetime64[M]').astype(int) % 12 + 1 for dti in df['timestamp'].values]
df['year'] = pd.DatetimeIndex(df['timestamp']).year
df = df.drop(['datetime', 'Unnamed: 0', '%s_%s_%s' % (variables_A[0], scenario, source_id) ], axis=1)

for source_id in models:
    for variable in variables_A:
        ff = os.path.join(data_dir, 'grid_data_extract', 'CMIP', '%s_%s_%s.csv' % (variable, scenario, source_id))
        dfi = pd.read_csv(ff)
        df['%s_%s_%s' % (variable, scenario, source_id)] = dfi['%s_%s_%s' % (variable, scenario, source_id)]
        if variable == 'tas':
            df['%s_%s_%s' % (variable, scenario, source_id)] = df['%s_%s_%s' % (variable, scenario, source_id)] - 273.15
        if variable == 'pr':
            df['%s_%s_%s' % (variable, scenario, source_id)] \
            = df['%s_%s_%s' % (variable, scenario, source_id)] * 60 * 60 * 24 / 1000 # [kg m-2 s-1] -> m


#df.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_%s_%s.csv' % (scenario, source_id)))
df.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_%s_models.csv' % (scenario)))
print(df.keys(), len(df.index))


df

scenario = scenarios[1]
ff = os.path.join(data_dir, 'grid_data_extract', 'CMIP', '%s_%s_%s.csv' % (variables_A[0], scenario, source_id))
df = pd.read_csv(ff)
#print(df.keys())
df['timestamp'] = [dt.strptime(dti, '%Y-%m-%d %H:%M:%S')
                                 for dti in df['time'].values]
df['month'] = [dti.astype('datetime64[M]').astype(int) % 12 + 1 for dti in df['timestamp'].values]
df['year'] = pd.DatetimeIndex(df['timestamp']).year
df = df.drop(['time', 'Unnamed: 0', '%s_%s_%s' % (variables_A[0], scenario, source_id) ], axis=1)

for source_id in models:
    for scenario in scenarios[1:]:
        for variable in variables_A:
            ff = os.path.join(data_dir, 'grid_data_extract', 'CMIP', '%s_%s_%s.csv' % (variable, scenario, source_id))
            dfi = pd.read_csv(ff)
            df['%s_%s_%s' % (variable, scenario, source_id)] = dfi['%s_%s_%s' % (variable, scenario, source_id)]
            if variable == 'tas':
                df['%s_%s_%s' % (variable, scenario, source_id)] = df['%s_%s_%s' % (variable, scenario, source_id)] - 273.15
            if variable == 'pr':
                df['%s_%s_%s' % (variable, scenario, source_id)] \
                = df['%s_%s_%s' % (variable, scenario, source_id)] * 60 * 60 * 24 / 1000 # [kg m-2 s-1] -> m
#df.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_proj_%s.csv' % (source_id))) 
df.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_proj_models.csv')) 

print(df.keys(), len(df.index))

df.head(5)

# Process growing season annual timeseries

gs_df = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'growing_season_locs_b.csv'))

gs_df

#data_combined = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_proj_%s.csv' % (source_id)))
data_combined = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_proj_models.csv'))

data_combines = data_combined.drop(['Unnamed: 0',], axis=1)

keys = [k for k in data_combined.keys() if k not in ['Lon', 'Lat', 'sp', 'year', 'month', 'timestamp']]

data_ym = data_combined.groupby(['Lon', 'Lat', 'sp', 'year']).mean()
data_ym = data_ym.drop(['month',], axis=1)
v_gs = []
v_gs_p = []
for lon, lat, sp, year in data_ym.index:
    
    gsi = gs_df[(np.round(gs_df['Lon'], 4) == np.round(lon, 4)) & (np.round(gs_df['Lat'], 4) == np.round(lat, 4))]
    if len(gsi.index)>=1:
        months_gs = gsi['months_gs'].values[0]
        months_gs = [np.int(m) for m in months_gs[1:-1].split(',')]
        peak_gs_month = gsi['peak_gs_month'].values[0]
        data_gs = data_combined[(data_combined['Lon']==lon) 
                       & (data_combined['Lat']==lat) 
                       & (data_combined['year']==year)]
        for m in range(1, 13):
            if m not in months_gs:
                data_gs = data_gs[data_gs['month']!=m]
        
        v_gs.append([np.mean(data_gs[kk]) for kk in keys])
        data_gs_peak = data_gs[data_gs['month']==peak_gs_month]
        v_gs_p.append([np.mean(data_gs_peak[kk]) for kk in keys])
    else:
        print(len(gsi.index)==0, lon, lat, year)
        v_gs.append([np.nan for kk in keys])
        v_gs_p.append([np.nan for kk in keys])

for v, kk in zip(zip(*v_gs), keys):
    data_ym['%s_gs' %kk] = v
for v, kk in zip(zip(*v_gs_p), keys):
    data_ym['%s_gs_peak' %kk] = v
data_ym = data_ym.reset_index()
data_ym.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_proj_models_annual.csv'))
#data_ym.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_proj_%s_annual.csv' % (source_id)))

data_combined = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_historical_models.csv'))

#data_combined = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'CMIP_historical_%s.csv' % (source_id)))
data_combines = data_combined.drop(['Unnamed: 0',], axis=1)

keys = [k for k in data_combined.keys() if k not in ['Lon', 'Lat', 'sp', 'year', 'month', 'timestamp']]

data_ym = data_combined.groupby(['Lon', 'Lat', 'sp', 'year']).mean()
data_ym = data_ym.drop(['month',], axis=1)
v_gs = []
v_gs_p = []
for lon, lat, sp, year in data_ym.index:
    
    gsi = gs_df[(np.round(gs_df['Lon'], 4) == np.round(lon, 4)) & (np.round(gs_df['Lat'], 4) == np.round(lat, 4))]
    if len(gsi.index)>=1:
        months_gs = gsi['months_gs'].values[0]
        months_gs = [np.int(m) for m in months_gs[1:-1].split(',')]
        peak_gs_month = gsi['peak_gs_month'].values[0]
        data_gs = data_combined[(data_combined['Lon']==lon) 
                       & (data_combined['Lat']==lat) 
                       & (data_combined['year']==year)]
        for m in range(1, 13):
            if m not in months_gs:
                data_gs = data_gs[data_gs['month']!=m]
        
        v_gs.append([np.mean(data_gs[kk]) for kk in keys])
        data_gs_peak = data_gs[data_gs['month']==peak_gs_month]
        v_gs_p.append([np.mean(data_gs_peak[kk]) for kk in keys])
    else:
        print(len(gsi.index)==0, lon, lat, year)
        v_gs.append([np.nan for kk in keys])
        v_gs_p.append([np.nan for kk in keys])

for v, kk in zip(zip(*v_gs), keys):
    data_ym['%s_gs' %kk] = v
for v, kk in zip(zip(*v_gs_p), keys):
    data_ym['%s_gs_peak' %kk] = v
data_ym = data_ym.reset_index()
#data_ym.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_%s_annual.csv' % (source_id)))
data_ym.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_models_annual.csv'))

data_ym

data_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'co2s_ssp245_GFDL-ESM4.csv'))
data_co2['timestamp'] = [dt.strptime(dti, '%Y-%m-%d %H:%M:%S')
                                 for dti in data_co2['datetime'].values]
data_co2['year'] = pd.DatetimeIndex(data_co2['timestamp']).year
data_co2 = data_co2.drop(['datetime', 'Unnamed: 0', 'CO2_ssp245', 'timestamp'], axis=1)

for scenario in scenarios[1:]:
    data_co2i = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'co2s_%s_GFDL-ESM4.csv' %scenario))
    data_co2['CO2_%s' % scenario] = data_co2i['CO2_%s' % scenario] * 10 ** (-6)
    
data_co2.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_proj_co2.csv'))
data_co2

scenario = 'historical'
data_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP', 'co2s_historical_GFDL-ESM4.csv'))
data_co2['timestamp'] = [dt.strptime(dti, '%Y-%m-%d %H:%M:%S')
                                 for dti in data_co2['datetime'].values]
data_co2['year'] = pd.DatetimeIndex(data_co2['timestamp']).year
data_co2 = data_co2.drop(['datetime', 'Unnamed: 0',  'timestamp'], axis=1)
data_co2['CO2_%s' % scenario] = data_co2['CO2_%s' % scenario] * 10 ** (-6)

data_co2.to_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_co2.csv'))
data_co2
