==> _data_management/4b_combine_model_inputs_LT_mean_outer_CMIP.ipynb <==
import os
import sys
import pandas as pd

import numpy as np
from datetime import datetime as dt
import calendar
import matplotlib.pyplot as plt

%load_ext autoreload
%autoreload 2

import warnings
warnings.filterwarnings('ignore')

data_dir = '/Users/maoyabassiouni/Documents/DATA/Projects/OLNI'

def cal_vpd(air_temp, rh):
    def __esat(air_temp):
        return 610.8 * np.exp((17.27 * air_temp) / (air_temp + 237.3))
    return __esat(air_temp) * (1 - rh/100)

# sites

df_ts_ = pd.read_csv(os.path.join(data_dir ,'model_inputs', 'df_model_all_annual.csv'))
df_ts_ = df_ts_[(df_ts_['yrange']>=10) & (df_ts_['count']>=6) 
                & (df_ts_['MODIS_Lai'].notna()) & (df_ts_['SATPSI'].notna())
                & (df_ts_['LNC'].notna())]
df_ts_ = df_ts_[(df_ts_['year']>=1995) & (df_ts_['year']<=2016)]

locations = list(df_ts_.groupby(['Lon', 'Lat', 'sp']).mean().index)
len(locations)

df_all = pd.read_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_LT.csv'))
df_all = df_all.set_index(['Lon', 'Lat', 'sp'])
df_all = df_all.loc[locations]
df_all

# Combine variables long-term mean

df_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_co2.csv'))
df_co2 = df_co2.drop([ 'Unnamed: 0'], axis=1)
df_co2 = df_co2[(df_co2['year']>=1995)]
df_all['CO2_historical_LTavg'] = np.mean(df_co2['CO2_historical'])
df_all['CO2_LTavg'] = np.mean(df_all['CO2'])
df_all = df_all.drop(['CO2'], axis=1)
df_all[['CO2_LTavg', 'CO2_historical_LTavg']]

df_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_proj_co2.csv'))
df_co2 = df_co2.drop([ 'Unnamed: 0'], axis=1)
df_all['CO2_ssp245_LTavg'] = np.mean(df_co2['CO2_ssp245'])
df_all['CO2_ssp585_LTavg'] = np.mean(df_co2['CO2_ssp585'])
df_all[['CO2_historical_LTavg', 'CO2_ssp245_LTavg',  'CO2_ssp585_LTavg']]

models = ['ACCESS-CM2',
          'CanESM5',
          'CESM2',
          'CNRM-CM6-1',
          'EC-Earth3-Veg-LR',
          'FGOALS-g3',
          'GFDL-CM4',
          'GISS-E2-1-G',
          'IITM-ESM',
          'INM-CM5-0',
          'IPSL-CM6A-LR',
          'KIOST-ESM',
          'MIROC6',
          'MPI-ESM1-2-LR',
          'MRI-ESM2-0',
          'NorESM2-MM',
          'UKESM1-0-LL']

data_C = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_models_annual.csv'))
data_C = data_C[(data_C['year']>=1995)]
data_C['Lat'] = np.round(data_C['Lat'], 4)
data_C['Lon'] = np.round(data_C['Lon'], 4)
data_C = data_C.groupby(['Lon', 'Lat', 'sp']).mean()
data_C = data_C.drop(['year', ], axis=1)
for k in data_C.keys():
    if k.startswith('Unnamed'):
        data_C = data_C.drop([k, ], axis=1)
keys = data_C.keys()
for k in keys:
    data_C['%s_LTavg'%k] = data_C[k]

for source_id in models:
    data_C['vpd_historical_%s_gs_LTavg' % (source_id)] = cal_vpd(data_C['tas_historical_%s_gs_LTavg' % (source_id)], 
                                                                 data_C['hurs_historical_%s_gs_LTavg' % (source_id)])
    data_C['par_historical_%s_gs_LTavg' % (source_id)] = data_C['rsds_historical_%s_gs_LTavg' % (source_id)] * 10 ** (-6) * 2.04

data_C = data_C.drop(keys, axis=1)

df_all = pd.merge(df_all, data_C, on =['Lat', 'Lon', 'sp'], how='outer')
df_all = df_all.drop(['Unnamed: 0', 'index'], axis=1)
df_all

data_C

### all model data

df_all.to_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_LT_CMIP.csv'))
