==> _data_management/5_combine_model_inputs_annual.ipynb <==
import os
import sys
import pandas as pd

import numpy as np
from datetime import datetime as dt
import calendar
import matplotlib.pyplot as plt

%load_ext autoreload
%autoreload 2

import warnings
warnings.filterwarnings('ignore')

data_dir = '/Users/maoyabassiouni/Documents/DATA/Projects/OLNI'

# Load LT data

df_all = pd.read_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_LT.csv'))
df_all = df_all.drop([ 'Unnamed: 0', 'index'], axis=1)
df_all['Lat'] = np.round(df_all['Lat'], 4)
df_all['Lon'] = np.round(df_all['Lon'], 4)
df_all = df_all.drop(['CO2',], axis=1)

df_all

# Join annual data

data_ERA5 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'era5_annual.csv'))
data_ERA5['Lat'] = np.round(data_ERA5['Lat'], 4)
data_ERA5['Lon'] = np.round(data_ERA5['Lon'], 4)
data_ERA5 = data_ERA5.groupby(['Lon', 'Lat', 'year']).mean()
data_ERA5 = data_ERA5.drop(['Unnamed: 0'], axis=1)

data_ERA5['tmp'] = data_ERA5['mean_2m_air_temperature']
data_ERA5['tmp_gs'] = data_ERA5['mean_2m_air_temperature_gs']
data_ERA5['tmp_gs_peak'] = data_ERA5['mean_2m_air_temperature_gs_peak']

data_ERA5['pre'] = data_ERA5['total_precipitation']
data_ERA5['pre_gs'] = data_ERA5['total_precipitation_gs']
data_ERA5['pre_gs_peak'] = data_ERA5['total_precipitation_gs_peak']

data_ERA5['pa'] = data_ERA5['surface_pressure']
data_ERA5['pa_gs'] = data_ERA5['surface_pressure_gs']
data_ERA5['pa_gs_peak'] = data_ERA5['surface_pressure_gs_peak']

data_ERA5 = data_ERA5.drop(['mean_2m_air_temperature', 'mean_2m_air_temperature_gs', 'mean_2m_air_temperature_gs_peak',
                           'total_precipitation', 'total_precipitation_gs', 'total_precipitation_gs_peak',
                           'surface_pressure', 'surface_pressure_gs', 'surface_pressure_gs_peak',], 
                           axis=1)

keys = data_ERA5.keys()
for kk in keys:
    data_ERA5['ERA5_%s_yi' % kk] = data_ERA5[kk]
    data_ERA5 = data_ERA5.drop(['%s' % kk,], axis=1)

data_ERA5 = data_ERA5.reset_index()
df_all = df_all.reset_index()
df_all = pd.merge(df_all, data_ERA5, on=['Lat', 'Lon', ], how='outer')
df_all = df_all.drop(['index'], axis=1)

print(len(data_ERA5.groupby(['Lon', 'Lat']).count().index))
print(len(data_ERA5.groupby(['Lon', 'Lat', 'year']).count().index)
      / len(data_ERA5.groupby(['Lon', 'Lat']).count().index))
print(len(df_all.groupby(['Lon', 'Lat', 'sp']).count().index))
print(keys)
df_all

data_TC = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'tc_annual.csv'))
data_TC['Lat'] = np.round(data_TC['Lat'], 4)
data_TC['Lon'] = np.round(data_TC['Lon'], 4)
data_TC = data_TC.groupby(['Lon', 'Lat', 'year']).mean()
data_TC = data_TC.drop(['Unnamed: 0'], axis=1)

data_TC['pre'] = data_TC['pr']
data_TC['pre_gs'] = data_TC['pr_gs']
data_TC['pre_gs_peak'] = data_TC['pr_gs_peak']
data_TC = data_TC.drop(['pr', 'pr_gs', 'pr_gs_peak'], axis=1)

data_TC['tmp'] = (data_TC['tmmn'] + data_TC['tmmx']) / 2
data_TC['tmp_gs'] = (data_TC['tmmn_gs'] + data_TC['tmmx_gs']) / 2
data_TC['tmp_gs_peak'] = (data_TC['tmmn_gs_peak'] + data_TC['tmmx_gs_peak']) / 2


data_TC['par'] = data_TC['srad'] * 10 ** (-6) * 2.04
data_TC['par_gs'] = data_TC['srad_gs'] * 10 ** (-6) * 2.04
data_TC['par_gs_peak'] = data_TC['srad_gs_peak'] * 10 ** (-6) * 2.04

keys = data_TC.keys()
for kk in keys:
    data_TC['TC_%s_yi' % kk] = data_TC[kk]
    data_TC = data_TC.drop(['%s' % kk,], axis=1)

data_TC = data_TC.reset_index()
df_all = df_all.reset_index()
df_all = pd.merge(df_all, data_TC, on=['Lat', 'Lon', 'year'], how='outer')
df_all = df_all.drop(['index'], axis=1)

print(len(data_TC.groupby(['Lon', 'Lat']).count().index))
print(len(df_all.groupby(['Lon', 'Lat', 'sp']).count().index))
print(keys)
df_all

df_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'co2_annual.csv'))
df_co2 = df_co2.drop([ 'Unnamed: 0'], axis=1)
df_co2 = df_co2[(df_co2['year']>=1990) & (df_co2['year']<=2016)]

df_co2 = df_co2.reset_index()
df_all = pd.merge(df_co2, df_all, on=['year', ], how='outer')
df_all

df_all = df_all.drop(['index'], axis=1)
print(len(df_all.groupby(['Lon', 'Lat', 'sp']).count().index))
df_all['CO2'] = df_all['CO2'] * 10 ** -6
df_all

### add obs data

dfR_icp = pyreadr.read_r(os.path.join(data_dir, 'raw', 'foliarNutrientsEnv_R', 'icp_trends.Rdata'))
dfR_icp = pd.DataFrame(dfR_icp['icp.trends'])
dfR_icp[['lat', 'lon', 'year', 'code_tree_species', 'n']].dropna()
dfR_icp['Lat'] = np.round(dfR_icp['lat'], 4)
dfR_icp['Lon'] = np.round(dfR_icp['lon'], 4)
dfR_icp['sp'] = dfR_icp['code_tree_species']
dfR_icp['LNC'] = dfR_icp['n']
dfR_icp = dfR_icp[['Lat', 'Lon', 'year', 'sp', 'LNC']]
dfR_icp = dfR_icp[(dfR_icp['LNC']>0) & (dfR_icp['year']>=1990) & (dfR_icp['year']<=2016)]

df_locs = pd.read_csv(os.path.join(data_dir, 'sites', 'all_icp_sites.csv'))
df_locs['Lat'] = np.round(df_locs['lat'], 4)
df_locs['Lon'] = np.round(df_locs['lon'], 4)
df_locs['sp'] = np.round(df_locs['code_tree_species'], 0)
df_locs = df_locs[['Lat', 'Lon', 'sp']]
df_locs = [list(arr) for arr in df_locs[['Lat', 'Lon', 'sp']].values]
print(len(df_locs))
df_locs

for ix, dfi in dfR_icp.iterrows():
    if list(dfi[['Lat', 'Lon', 'sp']].values) not in df_locs:
        dfR_icp = dfR_icp.drop(index=ix)
print(len(dfR_icp.groupby(['Lon', 'Lat', 'sp']).count().index))

df_all['sp'] = round(df_all['sp'], 0)
df_all = pd.merge(df_all, dfR_icp, on=['Lat', 'Lon', 'year', 'sp'], how='outer')
print(len(df_all.groupby(['Lon', 'Lat', 'sp']).count().index))
df_all

### all model data

df_all = df_all.groupby(['Lat', 'Lon', 'year', 'sp']).mean()
df_all.to_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_annual.csv'))

len(df_all.index)/(2016 - 1990+1)
