==> _data_management/5b_combine_model_inputs_annual_CMIP.ipynb <==
import os
import sys
import pandas as pd

import numpy as np
from datetime import datetime as dt
import calendar
import matplotlib.pyplot as plt

%load_ext autoreload
%autoreload 2

import warnings
warnings.filterwarnings('ignore')

data_dir = '/Users/maoyabassiouni/Documents/DATA/Projects/OLNI'

def cal_vpd(air_temp, rh):
    def __esat(air_temp):
        return 610.8 * np.exp((17.27 * air_temp) / (air_temp + 237.3))
    return __esat(air_temp) * (1 - rh/100)

def f_temp(air_temp, pa):
    a =  np.exp((17.27 * air_temp) / (air_temp + 237.3))
    g = 0.665 * pa
    return a / (a + g)

models = ['ACCESS-CM2',
          'CanESM5',
          'CESM2',
          'CNRM-CM6-1',
          'EC-Earth3-Veg-LR',
          'FGOALS-g3',
          'GFDL-CM4',
          'GISS-E2-1-G',
          'IITM-ESM',
          'INM-CM5-0',
          'IPSL-CM6A-LR',
          'KIOST-ESM',
          'MIROC6',
          'MPI-ESM1-2-LR',
          'MRI-ESM2-0',
          'NorESM2-MM',
          'UKESM1-0-LL']

df_all = pd.read_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_LT_CMIP.csv'))
df_all['Lat'] = np.round(df_all['Lat'], 4)
df_all['Lon'] = np.round(df_all['Lon'], 4)

data_CMIP_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_co2.csv'))
data_CMIP_co2 = data_CMIP_co2.drop(['Unnamed: 0',], axis=1)

data_CMIP = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_models_annual.csv'))
#data_CMIP = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_historical_CESM2_annual.csv'))
data_CMIP['Lat'] = np.round(data_CMIP['Lat'], 4)
data_CMIP['Lon'] = np.round(data_CMIP['Lon'], 4)
data_CMIP = data_CMIP.groupby(['Lon', 'Lat', 'sp', 'year']).mean()
for k in data_CMIP.keys():
    if k.startswith('Unnamed'):
        data_CMIP = data_CMIP.drop([k,], axis=1)

        
data_CMIP = data_CMIP.reset_index()
df_all = df_all.reset_index()
df_all = pd.merge(df_all, data_CMIP, on=['Lat', 'Lon', 'sp'], how='outer')
df_all = df_all.drop(['index'], axis=1)

co2 = [data_CMIP_co2[data_CMIP_co2['year']==yi]['CO2_historical'].values[0] for yi in df_all['year']]
df_all['CO2_historical'] = co2

#conversions.... 
for source_id in models:
    for tt in ['gs_LTavg', 'gs']:
        df_all['par_historical_%s_%s' % (source_id, tt)] = df_all['rsds_historical_%s_%s' % (source_id, tt)] * 10 ** (-6) * 2.04
        df_all['vpd_historical_%s_%s' % (source_id, tt)] = cal_vpd(df_all['tas_historical_%s_%s' % (source_id, tt)], df_all['hurs_historical_%s_%s' % (source_id, tt)])

        df_all['rf_alpha_historical_%s_%s' % (source_id, tt)] = df_all['pr_historical_%s_%s' % (source_id, tt)]\
                                                     /  df_all['ERA5_rf_lambda_gs_LTavg']
        df_all['rf_lambda_historical_%s_%s' % (source_id, tt)] = df_all['ERA5_rf_lambda_gs_LTavg']
        df_all['pet_historical_%s_%s' % (source_id, tt)] = df_all['TC_pet_gs_LTavg'] \
                                                * df_all['rsds_historical_%s_%s' % (source_id, tt)] / df_all['TC_srad_gs_LTavg'] \
                                                * f_temp(df_all['tas_historical_%s_%s' % (source_id, tt)], df_all['elev_pa']) \
                                                / f_temp(df_all['TC_tmmn_gs_LTavg'], df_all['elev_pa'])

df_all.to_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_annual_CMIP_historical.csv'))

df_all

list(data_CMIP.keys())

df_all = pd.read_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_LT_CMIP.csv'))
df_all['Lat'] = np.round(df_all['Lat'], 4)
df_all['Lon'] = np.round(df_all['Lon'], 4)

data_CMIP_co2 = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_proj_co2.csv'))
data_CMIP_co2 = data_CMIP_co2.drop(['Unnamed: 0',], axis=1)

data_CMIP = pd.read_csv(os.path.join(data_dir, 'grid_data_extract', 'CMIP_proj_models_annual.csv'))
data_CMIP['Lat'] = np.round(data_CMIP['Lat'], 4)
data_CMIP['Lon'] = np.round(data_CMIP['Lon'], 4)
data_CMIP = data_CMIP.groupby(['Lon', 'Lat', 'sp', 'year']).mean()
for k in data_CMIP.keys():
    if k.startswith('Unnamed'):
        data_CMIP = data_CMIP.drop([k,], axis=1)

        
data_CMIP = data_CMIP.reset_index()
df_all = df_all.reset_index()
df_all = pd.merge(df_all, data_CMIP, on=['Lat', 'Lon', 'sp'], how='outer')
df_all = df_all.drop(['index'], axis=1)

for ck in ['CO2_ssp245', 'CO2_ssp585']:
    co2 = [data_CMIP_co2[data_CMIP_co2['year']==yi][ck].values[0] for yi in df_all['year']]
    df_all[ck] = co2

#conversions.... 
for source_id in models:
    for sc in ['ssp245', 'ssp585']:
        for tt in [ 'gs',]:
            df_all['par_%s_%s_%s' % (sc, source_id, tt)] = df_all['rsds_%s_%s_%s' % (sc, source_id, tt)] * 10 ** (-6) * 2.04
            df_all['vpd_%s_%s_%s' % (sc, source_id, tt)] = cal_vpd(df_all['tas_%s_%s_%s' % (sc, source_id, tt)], df_all['hurs_%s_%s_%s' % (sc, source_id, tt)])

            df_all['rf_alpha_%s_%s_%s' % (sc, source_id, tt)] = df_all['pr_%s_%s_%s' % (sc, source_id, tt)]\
                                                         /  df_all['ERA5_rf_lambda_gs_LTavg']
            df_all['rf_lambda_%s_%s_%s' % (sc, source_id, tt)] = df_all['ERA5_rf_lambda_gs_LTavg']
            df_all['pet_%s_%s_%s' % (sc, source_id, tt)] = df_all['TC_pet_gs_LTavg'] \
                                                    * df_all['rsds_%s_%s_%s' % (sc, source_id, tt)] / df_all['TC_srad_gs_LTavg'] \
                                                    * f_temp(df_all['tas_%s_%s_%s' % (sc, source_id, tt)], df_all['elev_pa']) \
                                                / f_temp(df_all['TC_tmmn_gs_LTavg'], df_all['elev_pa'])

df_all.to_csv(os.path.join(data_dir, 'model_inputs', 'df_model_all_annual_CMIP_proj.csv'))

df_all

list(df_all.keys())
